+++ Plugin.cs
@@ -36,8 +36,7 @@ namespace Olympus;
 
 public sealed class Plugin : IDalamudPlugin
 {
-    public const string PluginVersion = "4.10.30";
-
+    public const string PluginVersion = "4.10.31";
     private const string CommandName = "/olympus";
 
     private readonly IDalamudPluginInterface pluginInterface;